+++ environment/providers/secretdir/secretdir_test.go
@@ -10,6 +10,21 @@ import (
 	"k8s.io/apimachinery/pkg/util/rand"
 )
 
+// Write regular and symlink file for given key-val pair.
+func writeParam(t *testing.T, path, key, val string) error {
+	filename := rand.String(10)
+	keyPath := filepath.Join(path, key)
+	dataPath := filepath.Join(path, filename)
+
+	if err := os.WriteFile(dataPath, []byte(val), 0o600); err != nil {
+		return err
+	}
+	if err := os.Symlink(filename, keyPath); err != nil {
+		return err
+	}
+	return nil
+}
+
 func TestBasicAuth(t *testing.T) {
 	g := NewWithT(t)
 	path, err := os.MkdirTemp(".", "basic-auth")
@@ -25,9 +40,7 @@ func TestBasicAuth(t *testing.T) {
 	user := rand.String(10)
 	passwd := rand.String(10)
 	key := fmt.Sprintf("%s:9440:%s:%s", ip, user, passwd)
-	g.Expect(os.WriteFile(filepath.Join(path, secretKeyName),
-		[]byte(key),
-		0o600)).To(Succeed())
+	g.Expect(writeParam(t, path, secretKeyName, key)).To(Succeed())
 	t.Log("Key", key)
 
 	prov := &provider{}
@@ -52,12 +65,9 @@ func TestTLSAuth(t *testing.T) {
 	cert := rand.String(512)
 	t.Logf("Temporary directory %s", path)
 	os.Setenv(envSecretDir, path)
-	g.Expect(os.WriteFile(filepath.Join(path, secretKeyCertName),
-		[]byte(cert),
-		0o600)).To(Succeed())
-	g.Expect(os.WriteFile(filepath.Join(path, secretKeyCertEndpoint),
-		[]byte(fmt.Sprintf("%s:9440", ip)),
-		0o600)).To(Succeed())
+	g.Expect(writeParam(t, path, secretKeyCertName, cert)).To(Succeed())
+	g.Expect(writeParam(t, path, secretKeyCertEndpoint,
+		fmt.Sprintf("%s:9440", ip))).To(Succeed())
 
 	prov := &provider{}
 	me, err := prov.GetManagementEndpoint(nil)